make registry dependencies available in 5.7
motivation: make experimental use of registry easier

changes:
* make registry based dependencies available in 5.7
* move the relevant tests to the 5.7 suite
