+++ Sources/PackageDescription/PackageDependency.swift
@@ -495,7 +495,7 @@ extension Package.Dependency {
     /// - Parameters:
     ///     - id: The identity of the package.
     ///     - version: The minimum version requirement.
-    @available(_PackageDescription, introduced: 999)
+    @available(_PackageDescription, introduced: 5.7)
     public static func package(
         id: String,
         from version: Version
@@ -519,7 +519,7 @@ extension Package.Dependency {
     /// - Parameters:
     ///     - id: The identity of the package.
     ///     - version: The minimum version requirement.
-    @available(_PackageDescription, introduced: 999)
+    @available(_PackageDescription, introduced: 5.7)
     public static func package(
         id: String,
         exact version: Version
@@ -548,7 +548,7 @@ extension Package.Dependency {
     /// - Parameters:
     ///     - id: The identity of the package.
     ///     - range: The custom version range requirement.
-    @available(_PackageDescription, introduced: 999)
+    @available(_PackageDescription, introduced: 5.7)
     public static func package(
         id: String,
         _ range: Range<Version>
@@ -567,7 +567,7 @@ extension Package.Dependency {
     /// - Parameters:
     ///     - id: The identity of the package.
     ///     - range: The closed version range requirement.
-    @available(_PackageDescription, introduced: 999)
+    @available(_PackageDescription, introduced: 5.7)
     public static func package(
         id: String,
         _ range: ClosedRange<Version>
@@ -582,7 +582,7 @@ extension Package.Dependency {
     }
 
     // intentionally private to hide enum detail
-    @available(_PackageDescription, introduced: 999)
+    @available(_PackageDescription, introduced: 5.7)
     private static func package(
         id: String,
         requirement: Package.Dependency.RegistryRequirement
