+++ Tests/PackageLoadingTests/PD_5_7_LoadingTests .swift
@@ -22,6 +22,33 @@ class PackageDescription5_7LoadingTests: PackageDescriptionLoadingTests {
         .v5_7
     }
 
+    func testRegistryDependencies() throws {
+        let content = """
+            import PackageDescription
+            let package = Package(
+               name: "MyPackage",
+               dependencies: [
+                   .package(id: "x.foo", from: "1.1.1"),
+                   .package(id: "x.bar", exact: "1.1.1"),
+                   .package(id: "x.baz", .upToNextMajor(from: "1.1.1")),
+                   .package(id: "x.qux", .upToNextMinor(from: "1.1.1")),
+                   .package(id: "x.quux", "1.1.1" ..< "3.0.0"),
+               ]
+            )
+            """
+
+        let observability = ObservabilitySystem.makeForTesting()
+        let manifest = try loadManifest(content, observabilityScope: observability.topScope)
+        XCTAssertNoDiagnostics(observability.diagnostics)
+
+        let deps = Dictionary(uniqueKeysWithValues: manifest.dependencies.map{ ($0.identity.description, $0) })
+        XCTAssertEqual(deps["x.foo"], .registry(identity: "x.foo", requirement: .range("1.1.1" ..< "2.0.0")))
+        XCTAssertEqual(deps["x.bar"], .registry(identity: "x.bar", requirement: .exact("1.1.1")))
+        XCTAssertEqual(deps["x.baz"], .registry(identity: "x.baz", requirement: .range("1.1.1" ..< "2.0.0")))
+        XCTAssertEqual(deps["x.qux"], .registry(identity: "x.qux", requirement: .range("1.1.1" ..< "1.2.0")))
+        XCTAssertEqual(deps["x.quux"], .registry(identity: "x.quux", requirement: .range("1.1.1" ..< "3.0.0")))
+    }
+
     func testConditionalTargetDependencies() throws {
         let content = """
             import PackageDescription
